requirements.txt
xlrd==2.0.1
xlwt==1.3.0
PySide6==6.2.2.1
python-docx>=0.8.11
lxml~=4.9.1